refactor: 렌더링 최적화를 위한 컴포넌트 분리
- 상태 변수가 Timetable에 존재해서 Timetable 전체가 리렌더링 되는 문제 발생
- 이를 해결하기 위해 LectureBox를 감싸는 Container를 따로 선언
