+++ client/src/components/UI/molecules/Timetable/Timetable.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import Box from '@material-ui/core/Box';
-import { LectureGrid, LectureBox } from '@/components/UI/atoms';
+import { LectureGrid } from '@/components/UI/atoms';
+import { LectureBoxContainer } from '@/components/UI/molecules';
 import { makeStyles } from '@material-ui/core/styles';
-import { useReactiveVar } from '@apollo/client';
-import { useStores } from '@/stores';
 
 interface TimetableProps {
   row: number;
@@ -55,9 +54,7 @@ const useStyles = makeStyles((theme) => ({
 
 const Timetable = ({ row, containedSat }: TimetableProps): JSX.Element => {
   const classes = useStyles({ row, containedSat });
-  const { timeTableStore } = useStores();
-  const Lectures = useReactiveVar(timeTableStore.state.lectures);
-  const selectedTabIdx = useReactiveVar(timeTableStore.state.selectedTabIdx);
+
   const fillTableHeader = () => {
     const array = [];
     array.push(<Box className={classes.headerFirst} />);
@@ -89,22 +86,12 @@ const Timetable = ({ row, containedSat }: TimetableProps): JSX.Element => {
       return makeRow(index);
     });
   };
-  const fillTableByLectures = () => {
-    if (selectedTabIdx === 0) return <></>;
-    const lectureInfos = Lectures[selectedTabIdx - 1];
-    if (!lectureInfos) return <></>;
-    return lectureInfos.map((elem) => {
-      if (typeof elem.time === 'string') return <></>;
-      return elem.time.map((time) => {
-        return <LectureBox starttime={time.start} endtime={time.end} name={elem.name} division={elem.class} prof={elem.prof} />;
-      });
-    });
-  };
+
   return (
     <Box className={classes.root}>
       {fillTableHeader()}
       {fillTable()}
-      {fillTableByLectures()}
+      <LectureBoxContainer />
     </Box>
   );
 };
